fix：字段调整、模型文件列表代码规范调整

+++ qmodel-module-model/qmodel-module-model-biz/src/main/java/tech/qiantong/qmodel/module/model/dal/dataobject/cacl/ModelCaclDO.java
@@ -137,7 +137,7 @@ public class ModelCaclDO extends BaseEntity {
     /**
      * 判断是文件的id还是地址的id，0:文件，1：接口
      */
-    private Long addressType;
+    private Integer addressType;
 
     /**
      * 文件计算的历史方案id
